pyproject.toml
[project]
name = "risk-detector-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "caio-dev-py",email = "caio.dev.araujo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)"
]

[tool.poetry]
packages = [{include = "risk_detector_ai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
